refactor: split job policy helpers
